+++ src/jadhav_lab_to_nwb/rivera_and_shukla_2025/rivera_and_shukla_2025_insert_session.py
@@ -23,7 +23,7 @@ from spyglass.spikesorting.analysis.v1.unit_annotation import UnitAnnotation
 
 # Custom Table Imports
 sys.path.append(
-    "/Users/pauladkisson/Documents/CatalystNeuro/JadhavConv/jadhav-lab-to-nwb/src/jadhav_lab_to_nwb/olson_2024/spyglass_extensions"
+    "/Users/pauladkisson/Documents/CatalystNeuro/JadhavConv/jadhav-lab-to-nwb/src/jadhav_lab_to_nwb/rivera_and_shukla_2025/spyglass_extensions"
 )
 from epoch import TaskLEDs
